@mansona 2025-04-12
Now that we have https://github.com/embroider-build/embroider/pull/2450 you should be able to rebase and change the tsconfig target accordingly to have this package support the ESM that you need