@simonihmig 2025-04-17
> to have this package support the ESM

Trying [here](https://github.com/embroider-build/embroider/pull/2454), but not successful so far. Any help would be appreciated.