@NullVoxPopuli 2025-04-18
Gonna merge, since the above esm efforts are kind of a side quest, and this improvement is amazing! Thank you @simonihmig !